feat(frontend): Vitest の in-source testing を導入する
admin/viewer に vitest を追加し、import.meta.vitest の型補完を有効化する。
純関数テストを in-source へ移し、テスト専用ファイルと不要な export を減らす。
