+++ src/admin/astro.config.mjs
@@ -34,9 +34,14 @@ export default defineConfig({
   },
   vite: {
     plugins: [tailwindcss()],
+    define: {
+      // in-source テストブロックを本番ビルドから除去する
+      'import.meta.vitest': 'undefined',
+    },
     ssr: {
       noExternal: shouldBundleServerDependencies ? true : undefined,
     },
   },
+
   integrations: [solidJs()],
 });
