+++ src/admin/src/server/csrf.ts
@@ -12,3 +12,25 @@ export const isCsrfTokenMatch = (expected: string, actual: string): boolean => {
 
   return timingSafeEqual(hash(expected), hash(actual));
 };
+
+if (import.meta.vitest) {
+  const { describe, expect, it } = import.meta.vitest;
+
+  describe('isCsrfTokenMatch', () => {
+    it('一致するトークンで true を返す', () => {
+      expect(isCsrfTokenMatch('csrf-token-value', 'csrf-token-value')).toBe(true);
+    });
+
+    it('同じ長さの異なるトークンで false を返す', () => {
+      expect(isCsrfTokenMatch('csrf-token-aaaa', 'csrf-token-bbbb')).toBe(false);
+    });
+
+    it('長さの異なるトークンでも例外を投げず false を返す', () => {
+      expect(isCsrfTokenMatch('short', 'a-much-longer-csrf-token-value')).toBe(false);
+    });
+
+    it('空文字と非空文字で false を返す', () => {
+      expect(isCsrfTokenMatch('', 'non-empty')).toBe(false);
+    });
+  });
+}
