+++ src/pnpm-workspace.yaml
@@ -10,6 +10,7 @@ packages:
 catalog:
   typescript: 6.0.3
   "@types/bun": 1.3.14
+  vitest: 4.1.9
   wrangler: 4.105.0
 
 catalogs:
